gitpkgv.bbclass: cache GITPKGV result
gitpkgv runs the 'git rev-list | wc -l' several times when processing a
package using GITPKGV.  This takes ages for packages like the linux
kernel which has a) a large repository and b) lots of subpackages.

This patch caches the result of 'git rev-list' into the sources cache
directory and uses it on the next run.  Because collisions of the sha1
hash are very unlikely, the git revision is used directly as the key.

[this is a slightly fixed version of classic OE's
 8c61580bc303911cf77a16eefb661414fee637f4]

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>