meta-systemd: select recipes based on active collections
Recipes of inactivate layers must been BBMASK'ed to prevent build errors.
This is very unflexible and causes high maintenance effort for keeping
the project configuration in sync with the actual meta-systemd layout.

With this patch, the .bbappend files within the meta-* directories are
ignored unless the corresponding collection has been selected.  As it
relies on the order of BBFILES and BBFILE_COLLECTIONS expansion, a
sanity check has been added which aborts the build when assumptions do
not apply anymore.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>